rearranging postId route for edit route
